Rename notificators to notifiers
